% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetricsLog2ratioInternalMethods.R
\encoding{UTF-8}
\name{calculateOneLog2valueMetric}
\alias{calculateOneLog2valueMetric}
\title{Calculate metric using the log2ratio values between two samples.}
\usage{
calculateOneLog2valueMetric(
  sample01,
  sample02,
  bedExclusion,
  method,
  minThreshold
)
}
\arguments{
\item{sample01}{a \code{GRanges} which contains a collection of 
genomic ranges representing copy number events for the first sample. 
The \code{GRanges} must have a metadata column called '\code{log2ratio}' 
with the log2ratio value for each segment.}

\item{sample02}{a \code{GRanges} which contains a collection of 
genomic ranges representing copy number events for the second sample.
The \code{GRanges} must have a metadata column called '\code{log2ratio}' 
with the log2ratio value for each segment.}

\item{bedExclusion}{a \code{GRanges} TODO}

\item{method}{a \code{character} string representing the metric to be
used ('\code{weightedEuclideanDistance}').}

\item{minThreshold}{a single \code{numeric} setting the minimum value 
to consider two segments as different during the metric calculation. If the 
absolute difference is below or equal to threshold, the difference will be 
replaced by zero. Default: 0.2.}
}
\value{
a \code{numeric}, the value of the specified metric. If
the metric cannot be calculated, \code{NA} is returned.
}
\description{
Calculate a specific metric using the level of 
amplification/deletion, in log2 ratio,  between 
two samples.
}
\details{
The method calculates a specified metric using overlapping
regions between the samples. Only regions corresponding to the type
specified by user are used in the calculation of the metric. The strand of 
the regions is not taken into account while
calculating the metric.

The Sorensen metric is calculated by dividing twice the size of 
the intersection by the sum of the size of the two sets. If the sum of
the size of the two sets is zero; the value \code{NA} is
returned instead.
}
\examples{

## Load required package to generate the two samples
require(GenomicRanges)

## Generate two samples with log2value information as a metadata column
sample01 <- GRanges(seqnames = "chr1", 
    ranges =  IRanges(start = c(100, 201, 400), 
    end = c(200, 350, 500)), strand =  "*",
    log2ratio = c(1.1111, 2.2222, -0.9999))
sample02 <- GRanges(seqnames = "chr1", 
    ranges =  IRanges(start = c(150, 200, 450), 
    end = c(250, 350, 500)), strand =  "*",
    log2ratio = c(2.2121, 1.1212, -1.3939))

## Calculate weighted Euclidean distance 
CNVMetrics:::calculateOneLog2valueMetric(sample01, sample02,
    bedExclusion=NULL, method="weightedEuclideanDistance", 
    minThreshold=0.2)


}
\references{
Sørensen, Thorvald. n.d. “A Method of Establishing Groups of Equal 
Amplitude in Plant Sociology Based on Similarity of Species and Its 
Application to Analyses of the Vegetation on Danish Commons.” 
Biologiske Skrifter, no. 5: 1–34.
}
\author{
Astrid Deschênes
}
\keyword{internal}
