% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetricsInternalMethods.R
\name{calculateOverlapMetric}
\alias{calculateOverlapMetric}
\title{Calculate metric using overlapping amplified/deleted regions between 
two samples.}
\usage{
calculateOverlapMetric(sample01, sample02, method, type)
}
\arguments{
\item{sample01}{a \code{GRanges} TODO}

\item{sample02}{a \code{GRanges} TODO}

\item{method}{a string of \code{character} representing the metric to be
used.}

\item{type}{a string of \code{character} TODO}
}
\value{
a \code{list} of TODO
}
\description{
Calculate a specific metric using overlapping regions between
two samples.
}
\examples{

# TODO

}
\author{
Astrid Deschênes
}
\keyword{internal}
