% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetricsMethods.R
\name{calculateOverlapRegionsMetric}
\alias{calculateOverlapRegionsMetric}
\title{Calculate metric using overlapping amplified/deleted regions between 
samples.}
\usage{
calculateOverlapRegionsMetric(
  segmentData,
  method = c("sorensen", "szymkiewicz")
)
}
\arguments{
\item{segmentData}{a \code{GRangesList} that contains the segment 
information, including amplified/deleted status, from at least 2 samples.}

\item{method}{a \code{character} string representing the metric to be used. 
This should be (an unambiguous abbreviation of) one of "sorensen" or 
"szymkiewicz-simpson". Default: "sorensen".}
}
\value{
a \code{matrix} TODO
}
\description{
Calculate a specific metric using overlapping regions between
to samples. When more than 2 samples are present, the metric is calculated
for each sample pair.
}
\details{
Available distance measures are (written for two Granges x and y):
# TODO
}
\examples{

## Create a GRangesList object with 3 samples

require(GenomicRanges)

demo <- GRangesList()
demo[["sample01"]] <- GRanges(seqnames = "chr1", 
    ranges =  IRanges(start = c(1905048, 4554832, 31686841), 
    end = c(2004603, 4577608, 31695808)), strand =  "*",
    state = c("AMPLIFICATION", "AMPLIFICATION", "DELETION"))
demo[["sample02"]] <- GRanges(seqnames = "chr1", 
    ranges =  IRanges(start = c(1995066, 31611222), 
    end = c(2204505, 31689898)), strand =  c("-", "+"),
    state = c("AMPLIFICATION", "AMPLIFICATION"))
demo[["sample03"]] <- GRanges(seqnames = "chr1", 
    ranges =  IRanges(start = c(1906069, 4558838), 
    end = c(1909505, 4570601)), strand =  "*",
    state = c("AMPLIFICATION", "DELETION"))

calculateOverlapRegionsMetric(demo)


}
\author{
Astrid Deschênes, Pascal Belleau
}
