% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetricsInternalMethods.R
\name{calculateSzymkiewicz}
\alias{calculateSzymkiewicz}
\title{Calculate Szymkiewicz-Simpson metric using overlapping 
amplified/deleted regions between two samples.}
\usage{
calculateSzymkiewicz(sample01, sample02)
}
\arguments{
\item{sample01}{a \code{GRanges} TODO}

\item{sample02}{a \code{GRanges} TODO}
}
\value{
a \code{list} of TODO
}
\description{
Calculate Szymkiewicz-Simpson metric using overlapping 
regions between two samples.
}
\examples{

# TODO

require(GenomicRanges)

sample01 <- GRanges(seqnames = "chr1", 
    ranges =  IRanges(start = c(1905048, 4554832, 31686841), 
    end = c(2004603, 4577608, 31695808)), strand =  "*")
sample02 <- GRanges(seqnames = "chr1", 
    ranges =  IRanges(start = c(1995066, 31611222), 
    end = c(2204505, 31689898)), strand =  c("+", "-"))
    
CNVMetrics:::calculateSzymkiewicz(sample01, sample02)
    
}
\author{
Astrid Deschênes
}
\keyword{internal}
