% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetricsMethods.R
\name{calculateWeightedEuclideanDistance}
\alias{calculateWeightedEuclideanDistance}
\title{TODO}
\usage{
calculateWeightedEuclideanDistance(results)
}
\arguments{
\item{results}{TODO}
}
\value{
a \code{matrix} TODO
}
\description{
TODO
}
\examples{

# TODO


}
\author{
Astrid Deschenes, Pascal Belleau
}
