% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetrics.R
\docType{package}
\name{CNVMetrics-package}
\alias{CNVMetrics-package}
\alias{CNVMetrics}
\title{CNVMetrics: Copy number variant metrics}
\description{
This package implements metrics to facilite copy number variant 
comparison among samples and/or methods.
}
\author{
Astrid Deschênes and
Pascal Belleau

Maintainer:
Astrid Deschenes <adeschen@hotmail.com>
}
\keyword{package}
