% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetricsInternalMethods.R
\name{createSegments}
\alias{createSegments}
\title{TODO}
\usage{
createSegments(fileList, sourceList, bedExclusion)
}
\arguments{
\item{fileList}{a \code{list} of \code{GRanges}, the segments from multiple
files.}

\item{sourceList}{a \code{list}}

\item{bedExclusion}{a \code{GRanges}, the regions that must be
excluded from the analysis. Default: \code{NULL}.}
}
\value{
a \code{GRanges} containing the segment information from the file.
}
\description{
TODO
}
\examples{

# TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}
\keyword{internal}
