% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{is.CNVMetric}
\alias{is.CNVMetric}
\title{Is an object of class \code{CNVMetric}}
\usage{
\method{is}{CNVMetric}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{\ldots}{further arguments passed to or from other methods.}
}
\value{
a \code{logical}.
}
\description{
Functions to test inheritance relationships between an 
object and class \code{CNVMetric}.
}
