% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetricsMethods.R
\name{prepareInformation}
\alias{prepareInformation}
\title{Read segment files and exclusion file to generate a dataset
that can be used to calculate the metrics.}
\usage{
prepareInformation(segDirectory, chrInfo, bedExclusionFile = NULL,
  segmentWithHeader = FALSE)
}
\arguments{
\item{segDirectory}{a \code{character} string, the path to the directory
containing the segment files to compared. Only files with the ".seg" (or 
".SEG") extension will be used. At least 2 segment files are needed to
be able to used the metrics.}

\item{chrInfo}{a \code{Seqinfo} containing the name and the length of the
chromosomes to analyze. Only the chomosomes contained in this
\code{Seqinfo} will be analyzed.}

\item{bedExclusionFile}{a \code{character} string or \code{NULL}, the path 
to the BED file that contains the regions that must be excluded from the
analysis. If \code{NULL}, it is assume that there is no region of
exclusion for the calculation of the metrics. Default: \code{NULL}.}

\item{segmentWithHeader}{a \code{logical}, when \code{TRUE}, the
segment files have all a header that should not be imported. 
Default: \code{FALSE}.}
}
\value{
a \code{GRanges} containing the segment information from the file.
}
\description{
TODO
}
\details{
The supported BED format requires that the first three columns are: 
\itemize{
\item { 1. chrom - The name of the chromosome.}
\item { 2. chromStart - The starting position of the feature in the 
chromosome or scaffold. The first base in a chromosome is numbered 0.}
\item { 3. chromEnd - The ending position of the feature in the chromosome 
or scaffold. The chromEnd base is not included in the display of the 
feature. }}
}
\examples{

# TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}
