% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{print.CNVMetric}
\alias{print.CNVMetric}
\title{Print CNVMetric object}
\usage{
\method{print}{CNVMetric}(x, ...)
}
\arguments{
\item{x}{the output object from \code{calculateOverlapRegionsMetric}
function to be printed.}

\item{\ldots}{further arguments passed to or from other methods.}
}
\value{
the argument \code{x}.
}
\description{
Print a \code{CNVMetric} object and returns it
\code{invisibly}.
}
\seealso{
The default method \code{\link[base]{print.default}}.
}
