% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readSEGFile}
\alias{readSEGFile}
\title{Read segment file and return a \code{GRanges}}
\usage{
readSEGFile(filepath, uniqueTag, header = FALSE)
}
\arguments{
\item{filepath}{a \code{character} string, the path to the file to 
read. Alternatively can be a connection.}

\item{uniqueTag}{a \code{character} string, the unique name assigned to
the segment data associated to this file. Each imported file should
have a unique tag.}

\item{header}{a \code{logical}, when \code{TRUE}, the
file has a header that should not be imported. Default: \code{FALSE}.}
}
\value{
a \code{GRanges} containing the segment information from the file.
}
\description{
Read a segment file and extract the relevant
information to generate a \code{GRanges}.
}
\examples{

## Path to a demo segmentation file
segFile <- system.file("extdata", "demo01.seg", package = "CNVMetrics")

## This segment file has a header
## Load the content of the segment file
readSEGFile(filepath = segFile, uniqueTag = "demo01", header = TRUE)

}
\author{
Astrid Deschenes, Pascal Belleau
}
