% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetricsInternalMethods.R
\name{validateCalculateOverlapMetricParameters}
\alias{validateCalculateOverlapMetricParameters}
\title{Parameters validation for the \code{\link{calculateOverlapMetric}} 
function}
\usage{
validateCalculateOverlapMetricParameters(states, nJobs)
}
\arguments{
\item{states}{a \code{vector} of \code{character} string with at least one
entry. The strings are representing the states that will be analyzed.}

\item{nJobs}{a single positive \code{integer} specifying the number of 
worker jobs to create in case of distributed computation.}
}
\value{
\code{0}.
}
\description{
Validation of all parameters needed by the public
\code{\link{calculateOverlapMetric}} function.
}
\details{
#' @title calculate the regressed values for all segment file.
#' 
#' @description Use the linear regression model obtained for each paired of
#' segment files (current file versus reference) to calculate the regressed 
#' values for all segment files except for the reference file.
#' 
#' @param segmentData a \code{list} of that 
#' contains the segments from multiple files. The \code{list} is composed of 
#' those entries:
#' \itemize{
#' \item a \code{segment} that contains the \code{GRanges} with the segment
#' information.
#' \item a \code{regression} that contains the result of the paired 
#' regressions. The number of entries corresponds to the number of paired
#' segment files considering that the reference file is always the same. Each 
#' entry is a \code{list} composed of those entries:
#' \itemize{
#' \item a \code{x_used} with the name of the file that segments values
#' are used as x values in the regression model.
#' \item a \code{y_used} with the name of the file that segments values
#' are used as y values in the regression model. This is the reference file. 
#' It is the same for all paired regressions.
#' \item a \code{lm} that contains the regression model.
#' }
#' }
#' 
#' @return a \code{list} of that 
#' contains the segments from multiple files. The \code{list} is composed of 
#' those entries:
#' \itemize{
#' \item a \code{segments} that contains the \code{GRanges} with the segment
#' information and values from each segment file.
#' \item a \code{regression} that contains the result of the paired 
#' regressions.
#' \item a \code{regressedData} that contains the \code{GRanges} with the 
#' segment information and the regressed values for each segment file.
#' }
#' 
#' @examples
#'
#' # TODO
#' 
#' @author Astrid Deschênes, Pascal Belleau
#' @importFrom GenomicRanges elementMetadata elementMetadata<-
#' @importFrom stats predict
#' @keywords internal
calculateRegressedValues <- function(segmentData) {
    
    segments <- elementMetadata(segmentData$segments)

    #segmentData$regressed <- segments
    
    for (i in seq_len(length(segmentData$regression))) {
        lmData <- segmentData$regression[[i]][["lm"]]
        xName <- segmentData$regression[[i]][["x_used"]]
        tempVal <- data.frame(x=segments[, xName])
        segments[, xName] <- as.vector(predict(lmData, newdata = tempVal))
    }
    
    segmentData$regressedData <- segmentData$segments
    elementMetadata(segmentData$regressedData) <- segments
    
    return(segmentData)
}
}
\examples{


## Return zero as all parameters are valid
CNVMetrics:::validateCalculateOverlapMetricParameters(
    states="GAIN", nJobs=1)

}
\author{
Astrid Deschênes
}
\keyword{internal}
