% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetricsOverlapInternalMethods.R
\name{calculateOverlapMetric}
\alias{calculateOverlapMetric}
\title{Calculate metric using overlapping amplified/deleted regions between 
two samples.}
\usage{
calculateOverlapMetric(sample01, sample02, method, type)
}
\arguments{
\item{sample01}{a \code{GRanges} which contains a collection of 
genomic ranges representing copy number events for the first sample. 
The \code{GRanges} must have a metadata column called '\code{state}' with 
amplified regions identified as '\code{AMPLIFICATION}' and 
deleted regions identified as '\code{DELETION}'; regions with different 
identifications will not be used in the
calculation of the metric.}

\item{sample02}{a \code{GRanges} which contains a collection of 
genomic ranges representing copy number events for the second sample.}

\item{method}{a \code{character} string representing the metric to be
used ('\code{sorensen}' or '\code{szymkiewicz}'.}

\item{type}{a \code{character} string representing the type of 
copy number events to be used ('\code{AMPLIFICATION}' or '\code{DELETION}').}
}
\value{
a \code{numeric}, the value of the specified metric. If
the metric cannot be calculated, \code{NA} is returned.
}
\description{
Calculate a specific metric using overlapping 
amplified/deleted regions between two samples.
}
\details{
The method calculates a specified metric using overlapping
regions between the samples. Only regions corresponding to the type
specified by user are used in the calculation of the metric. The strand of 
the regions is not taken into account while
calculating the metric.

The Sorensen metric is calculated by dividing twice the size of 
the intersection by the sum of the size of the two sets. If the sum of
the size of the two sets is zero; the value \code{NA} is
returned instead. 

The Szymkiewicz-Simpson metric is calculated by dividing the size of 
the intersection by the smaller of the size of the two sets. If one sample
has a size of zero, the metric is not calculated; the value \code{NA} is
returned instead.
}
\examples{

## Load required package to generate the two samples
require(GenomicRanges)

## Generate two samples with identical sequence levels
sample01 <- GRanges(seqnames = "chr1", 
    ranges =  IRanges(start = c(100, 201, 400), 
    end = c(200, 350, 500)), strand =  "*",
    state = c("AMPLIFICATION", "AMPLIFICATION", "DELETION"))
sample02 <- GRanges(seqnames = "chr1", 
    ranges =  IRanges(start = c(150, 200, 450), 
    end = c(250, 350, 500)), strand =  "*",
    state = c("AMPLIFICATION", "DELETION", "DELETION"))

## Calculate Sorensen metric for the amplified regions   
CNVMetrics:::calculateOverlapMetric(sample01, sample02, method="sorensen",
    type="AMPLIFICATION")

## Calculate Szymkiewicz-Simpson metric for the amplified regions   
## Amplified regions of sample02 are a subset of the amplified 
## regions in sample01
CNVMetrics:::calculateOverlapMetric(sample01, sample02, method="szymkiewicz",
    type="AMPLIFICATION")

## Calculate Sorensen metric for the deleted regions   
CNVMetrics:::calculateOverlapMetric(sample01, sample02, method="sorensen",
    type="DELETION")
    
## Calculate Szymkiewicz-Simpson metric for the deleted regions    
CNVMetrics:::calculateOverlapMetric(sample01, sample02, method="szymkiewicz",
    type="DELETION")

}
\references{
Sørensen, Thorvald. n.d. “A Method of Establishing Groups of Equal 
Amplitude in Plant Sociology Based on Similarity of Species and Its 
Application to Analyses of the Vegetation on Danish Commons.” 
Biologiske Skrifter, no. 5: 1–34.

Vijaymeena, M. K, and Kavitha K. 2016. “A Survey on Similarity Measures in 
Text Mining.” Machine Learning and Applications: An International 
Journal 3 (1): 19–28. doi: \url{https://doi.org/10.5121/mlaij.2016.3103}
}
\author{
Astrid Deschênes
}
\keyword{internal}
