% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetricsMethods.R
\name{calculateWeightedEuclideanDistance}
\alias{calculateWeightedEuclideanDistance}
\title{Calculate Weighted Euclidean distance metric between samples.}
\usage{
calculateWeightedEuclideanDistance(results)
}
\arguments{
\item{results}{TODO}
}
\value{
a \code{matrix} TODO
}
\description{
The weighted euclidean distance metric correspond to the
euclidean distance multiplied by the natural logarithm of the number of 
bases of the segment. The final metric is the sum of the values obtained for
all segments that are not excluded of the analysis.
}
\examples{

# TODO


}
\author{
Astrid Deschenes, Pascal Belleau
}
