% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetricsMethods.R
\name{calculateWeightedEuclideanDistance}
\alias{calculateWeightedEuclideanDistance}
\title{Calculate Weighted Euclidean distance metric between samples.}
\usage{
calculateWeightedEuclideanDistance(segmentData, minThreshold = 0.2)
}
\arguments{
\item{segmentData}{a \code{list} marked as a \code{preMetricSegments} 
\code{class} that contains the segment information from at least 2 segment 
files.}

\item{minThreshold}{a single \code{numeric} setting the minimum value 
to consider two segments as different durant the metric calculation. If the 
absolute difference is bellow or equal to threshold, the value will be 
replaced by zero. Default: 0.2.}
}
\value{
a \code{matrix} TODO
}
\description{
The weighted Euclidean distance metric corresponds to the
euclidean distance between 2 samples multiplied by the natural logarithm 
of the number of bases of the analyzed segment. The final metric is the 
squared sum of the values obtained for all segments that are not 
excluded of the analysis.
}
\details{
The weighted euclidean distance is 
\eqn{(\sum((x_i - y_i)^2 * log(nbrBases_i))^0.5} 
where \code{x} and \code{y} are the
values of 2 samples for a specific segment \code{i} and \code{nbrBases} the 
number of bases of the segment \code{i}.
}
\examples{

# TODO


}
\author{
Astrid Deschenes, Pascal Belleau
}
