% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetricsInternalMethods.R
\name{createSegments}
\alias{createSegments}
\title{Generate common segments to enable calculation of metrics on 
multiple segment files.}
\usage{
createSegments(fileList, sourceList, bedExclusion)
}
\arguments{
\item{fileList}{a \code{list} of \code{GRanges}, the segments from multiple
files.}

\item{sourceList}{a \code{list}}

\item{bedExclusion}{a \code{GRanges}, the regions that must be
excluded from the analysis. Default: \code{NULL}.}
}
\value{
a \code{GRanges} containing the segment information from the file.
}
\description{
All segments are gathered together, including exclusion 
segments when specified, and a disjoin operation is done to create a 
collection of non-overlapping ranges. The ranges included in the exclusion
segments are marked as so to be removed from futur analysis.
}
\examples{

# TODO

}
\author{
Astrid Deschenes, Pascal Belleau
}
\keyword{internal}
