% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{is.CNVMetric}
\alias{is.CNVMetric}
\title{Is an object of class \code{CNVMetric}}
\usage{
\method{is}{CNVMetric}(object, class2)
}
\arguments{
\item{object}{an object.}

\item{class2}{a \code{character} string giving the name of the class.}
}
\value{
a \code{logical}.
}
\description{
Functions to test inheritance relationships between an
object and class \code{CNVMetric}.
}
