% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetricsMethods.R
\name{prepareInformation}
\alias{prepareInformation}
\title{Read segment files and their values and exclusion file to generate 
a dataset that can be used to calculate the metrics.}
\usage{
prepareInformation(
  segDirectory,
  chrInfo,
  bedExclusionFile = NULL,
  segmentWithHeader = FALSE
)
}
\arguments{
\item{segDirectory}{a \code{character} string, the path to the directory
containing the segment files to compare. Only files with the ".seg" (or 
".SEG") extension will be used. At least 2 segment files are needed to
be able to used the metrics.}

\item{chrInfo}{a \code{Seqinfo} containing the name and the length of the
chromosomes to analyze. Only the chromosomes contained in this
\code{Seqinfo} will be analyzed.}

\item{bedExclusionFile}{a \code{character} string or \code{NULL}, the path 
to the BED file that contains the regions that must be excluded from the
analysis. If \code{NULL}, it is assumed that there is no region of
exclusion for the calculation of the metrics. Default: \code{NULL}.}

\item{segmentWithHeader}{a \code{logical}, when \code{TRUE}, the
segment files have all a header that should not be imported. 
Default: \code{FALSE}.}
}
\value{
a \code{list} marked as a \code{preMetricSegments} \code{class}  
containing a \code{GRanges} with the segment information from all 
segment files.
}
\description{
The segments and their values are extracted from each file, 
with a "SEG" or "seg" extension, present in the 
specified directory. Each segment is associated to its original file by
having a source metadata field assigned. When specified, the segments
for a BED file are loaded and used to exclude regions from further analysis.
All segments are gathered together
and a disjoin operation is done to create a collection of non-overlapping 
ranges. Those ranges are going to be used to calculate the metrics.
}
\details{
The supported BED format requires that the first three columns are: 
\itemize{
\item { 1. chrom - The name of the chromosome.}
\item { 2. chromStart - The starting position of the feature in the 
chromosome or scaffold. The first base in a chromosome is numbered 0.}
\item { 3. chromEnd - The ending position of the feature in the chromosome 
or scaffold. The chromEnd base is not included in the display of the 
feature. }}
}
\examples{

# Path to the exclusion BED file
bedFile <- system.file("extdata", "mm10.blacklist.bed", 
    package = "CNVMetrics")

## Prepare information about the chromosomes
## The files are from mouse samples, chr1 to chr4
require(rtracklayer)

## Get the information for mouse genome mm10
## Limit the information to chromosomes 1 to 4
mouseInfo <- SeqinfoForUCSCGenome("mm10")  
mouseInfo <- mouseInfo[c("chr1", "chr2", "chr3", "chr4"),]
            
## Path to the directory containing the segmentation files
segDir <- system.file("extdata", package="CNVMetrics")

## TODO

prepareInformation(segDirectory = segDir, chrInfo = mouseInfo,
    bedExclusionFile = bedFile, segmentWithHeader = TRUE)

}
\author{
Astrid Deschenes, Pascal Belleau
}
