% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\encoding{UTF-8}
\name{processChr}
\alias{processChr}
\title{TODO}
\usage{
processChr(curSample, dfChr, chrCur)
}
\arguments{
\item{curSample}{a \code{GRangesList} that contains a collection of
genomic ranges representing copy number events, including amplified/deleted
status, from at least 1 samples. The sample must have a metadata column
called '\code{state}' with a state, in an character string format,
specified for each region (ex: DELETION, LOH, AMPLIFICATION, NEUTRAL, etc.).}

\item{dfChr}{data.frame}

\item{chrCur}{a \code{integer} which is corresponding to the number of simulation}
}
\value{
df TODO
}
\description{
TODO
}
\details{
TODO

metric value of 1 is only obtained when the two samples are identical.
}
\examples{
## TODO

}
\author{
Astrid Deschênes, Pascal Belleau
}
