% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVMetricsInternalMethods.R
\name{validateCalculateOverlapMetricParameters}
\alias{validateCalculateOverlapMetricParameters}
\title{Parameters validation for the \code{\link{calculateOverlapMetric}}
function}
\usage{
validateCalculateOverlapMetricParameters(states, nJobs)
}
\arguments{
\item{states}{a \code{vector} of \code{character} string with at least one
entry. The strings are representing the states that will be analyzed.}

\item{nJobs}{a single positive \code{integer} specifying the number of
worker jobs to create in case of distributed computation.}
}
\value{
\code{0}.
}
\description{
Validation of all parameters needed by the public
\code{\link{calculateOverlapMetric}} function.
}
\examples{


## Return zero as all parameters are valid
CNVMetrics:::validateCalculateOverlapMetricParameters(
    states="GAIN", nJobs=1)

}
\author{
Astrid Deschênes
}
\keyword{internal}
